backend/requirements.txt
fastapi
requests
python-dotenv
uvicorn
google-generativeai